requirements.txt
# Camera Capture Requirements

# NumPy (compatible with OpenCV 4.8.x)
numpy<2

# OpenCV for camera access and image processing
opencv-python==4.8.1.78

# AWS SDK for Python (will be used for S3 upload)
boto3==1.34.34

# Python dotenv for environment variable management
python-dotenv==1.0.0
openai
Flask==3.0.0
requests==2.31.0

# Realtime API & Web App
websockets
fastapi
uvicorn
python-multipart
pyaudio
